Split query classes into separate files.
