+++ flowmachine/tests/test_subscriber_event_count.py
@@ -53,67 +53,6 @@ def test_event_count_errors(kwarg):
         query = EventCount("2016-01-03", "2016-01-05", **{kwarg: "error"})
 
 
-@pytest.mark.parametrize(
-    "statistic,msisdn,want,level",
-    [
-        ("count", "Rzx9WE1QRqdEX2Gp", 16, {}),
-        ("sum", "LBlWd64rqnMGv7kY", 22, {}),
-        ("avg", "JZoaw2jzvK2QMKYX", 1.333_333, {}),
-        ("avg", "JZoaw2jzvK2QMKYX", 1.647_059, {"level": "admin3"}),
-        ("avg", "JZoaw2jzvK2QMKYX", 1.285_714_2, {"direction": "in"}),
-        ("max", "DELmRj9Vvl346G50", 4, {}),
-        ("min", "9vXy462Ej8V1kpWl", 1, {}),
-        ("stddev", "EkpjZe5z37W70QKA", 0.594_089, {}),
-        ("variance", "JNK7mk5G1Dy6M2Ya", 0.395_833, {}),
-    ],
-)
-def test_per_location_event_stats(get_dataframe, statistic, msisdn, want, level):
-    """ Test hand-picked PerLocationEventStats. """
-    query = PerLocationEventStats(
-        "2016-01-01", "2016-01-06", statistic=statistic, **level
-    )
-    df = get_dataframe(query).set_index("subscriber")
-    assert df.value[msisdn] == pytest.approx(want)
-
-
-@pytest.mark.parametrize("kwarg", ["direction", "statistic"])
-def test_per_location_event_stats_errors(kwarg):
-    """ Test ValueError is raised for non-compliant kwarg in PerLocationEventStats. """
-
-    with pytest.raises(ValueError):
-        query = PerLocationEventStats("2016-01-03", "2016-01-05", **{kwarg: "error"})
-
-
-@pytest.mark.parametrize(
-    "statistic,msisdn,want",
-    [
-        ("count", "Rzx9WE1QRqdEX2Gp", 2),
-        ("sum", "LBlWd64rqnMGv7kY", 16),
-        ("avg", "JZoaw2jzvK2QMKYX", 12.5),
-        ("max", "DELmRj9Vvl346G50", 14),
-        ("min", "9vXy462Ej8V1kpWl", 4),
-        ("median", "KXVqP6JyVDGzQa3b", 8),
-        ("stddev", "EkpjZe5z37W70QKA", 0),
-        ("variance", "JNK7mk5G1Dy6M2Ya", 2),
-    ],
-)
-def test_per_contact_event_stats(get_dataframe, statistic, msisdn, want):
-    """ Test hand-picked PerContactEventStats. """
-    query = PerContactEventStats(ContactBalance("2016-01-02", "2016-01-06"), statistic)
-    df = get_dataframe(query).set_index("subscriber")
-    assert df.value[msisdn] == pytest.approx(want)
-
-
-@pytest.mark.parametrize("kwarg", ["statistic"])
-def test_per_contact_event_stats_errors(kwarg):
-    """ Test ValueError is raised for non-compliant kwarg in PerContactEventStats. """
-
-    with pytest.raises(ValueError):
-        query = PerContactEventStats(
-            ContactBalance("2016-01-02", "2016-01-06"), **{kwarg: "error"}
-        )
-
-
 def test_directed_count_consistent(get_dataframe):
     """
     Test that directed count is consistent.
